+++ src/index.js
@@ -1,8 +1,15 @@
+/**
+ * Copyright (c) 2015 Hideki Shiro
+ */
+
 'use strict';
 
+var path = require('path');
 var fs = require('fs');
 var assert = require('assert');
 var opentype = require('opentype.js');
+
+const DEFAULT_FONT = path.join(__dirname, '../fonts/ipag.ttf');
 
 function toArrayBuffer(buffer) {
   var ab = new ArrayBuffer(buffer.length);
@@ -31,25 +38,38 @@ function cmdToSVG(cmd) {
 }
 
 export class TextToSVG {
-  constructor(url) {
-    let buf = fs.readFileSync(url);
+  constructor(file = DEFAULT_FONT) {
+    let buf = fs.readFileSync(file);
     let ab = toArrayBuffer(buf);
     this.font = opentype.parse(ab);
   }
 
-  getD(text, x, y, fontSize, options) {
-    let path = this.font.getPath(text, x, y, fontSize, options);
+  getD(text, options = {}) {
+    let x = options.x || 0;
+    let y = options.y || 0;
+    let fontSize = options.fontSize || 72;
+    let kerning = 'kerning' in options ? options.kerning : true;
+
+    let path = this.font.getPath(text, x, y, fontSize, {kerning});
     return path.commands.map(cmd => cmdToSVG(cmd)).join(' ');
   }
 
-  getPath(text, x, y, fontSize, options) {
-    let d = this.getD(text, x, y, fontSize, options);
-    return `<path d="${d}"/>`;
+  getPath(text, options = {}) {
+    options.attributes = options.attributes || {};
+
+    let attributes = Object.keys(options.attributes).map((key) => `${key}="${options.attributes[key]}"`).join(' ');
+    let d = this.getD(text, options);
+
+    if (attributes) {
+      return `<path ${attributes} d="${d}"/>`;
+    } else {
+      return `<path d="${d}"/>`;
+    }
   }
 
-  getSVG(text, x, y, fontSize, options) {
+  getSVG(text, options = {}) {
     let svg = '<svg xmlns="http://www.w3.org/2000/svg" xmlns:xlink="http://www.w3.org/1999/xlink">';
-    svg += this.getPath(text, x, y, fontSize, options);
+    svg += this.getPath(text, options);
     svg += '</svg>';
 
     return svg;