+++ CHANGELOG.md
@@ -1,8 +1,14 @@
 # CHANGELOG
 
+## 2.0.0
+
+ - Reform the API.
+ - Include default font.
+ - Remove the command line tool.
+
 ## 1.1.0
 
-- Add TextToSVG.getD(text, x, y, fontSize, options);
+ - Add TextToSVG.getD(text, x, y, fontSize, options);
 
 ## 1.0.0